requirements.txt
gitpython>=3.1.32
psycopg2>=2.9.7